Add and use function to generate entities, without base variables.
