+++ feature_bone/src/main/java/com/hfad/palamarchuksuperapp/feature/bone/ui/viewModels/PaymentPageState.kt
@@ -2,9 +2,9 @@ package com.hfad.palamarchuksuperapp.feature.bone.ui.viewModels
 
 import com.hfad.palamarchuksuperapp.feature.bone.domain.models.PaymentOrder
 import com.hfad.palamarchuksuperapp.feature.bone.domain.models.PaymentStatistic
-import com.hfad.palamarchuksuperapp.feature.bone.ui.screens.generateSamplePayments
+import com.hfad.palamarchuksuperapp.feature.bone.domain.models.generatePaymentOrderItems
 
 data class PaymentPageState(
-    val payments: List<PaymentOrder> = generateSamplePayments(), //TODO for test only
+    val payments: List<PaymentOrder> = generatePaymentOrderItems(), //TODO for test only
     val paymentStatistic: PaymentStatistic = PaymentStatistic(),
 )
