+++ feature_bone/src/main/java/com/hfad/palamarchuksuperapp/feature/bone/ui/viewModels/SalesPageState.kt
@@ -2,9 +2,9 @@ package com.hfad.palamarchuksuperapp.feature.bone.ui.viewModels
 
 import com.hfad.palamarchuksuperapp.feature.bone.domain.models.SaleOrder
 import com.hfad.palamarchuksuperapp.feature.bone.domain.models.SalesStatistics
-import com.hfad.palamarchuksuperapp.feature.bone.ui.composables.generateSampleProductSaleItems
+import com.hfad.palamarchuksuperapp.feature.bone.domain.models.generateSaleOrderItems
 
 data class SalesPageState(
-    val salesItems: List<SaleOrder> = generateSampleProductSaleItems(),
+    val salesItems: List<SaleOrder> = generateSaleOrderItems(),
     val salesStatistics: SalesStatistics = SalesStatistics(),
 )
